Develop

See merge request projetm2/backend!23

+++ src/Union.Backend.Model/Models/Criteria.cs
@@ -24,7 +24,7 @@ namespace Union.Backend.Model.Models
             set => RealLocationTime = value?.ToTimeSpan(); 
         }
         public int? Area { get; set; }
-        public double? Price { get; set; }
+        public float? Price { get; set; }
         public Orientation? Orientation { get; set; }
         public string TypeOfClay { get; set; }
         public bool? Equipments { get; set; }
